Created a file for UIAccessibility extension and fixed title for larger font size
