+++ VocabularyTrainer/Extensions/UIView+Extensions.swift
@@ -26,13 +26,3 @@ extension UIView {
                        completion: nil)
     }
 }
-
-extension UIAccessibility {
-    static func focusOn(_ object: Any?) {
-        if UIAccessibility.isVoiceOverRunning {
-            DispatchQueue.main.asyncAfter(deadline: .now()) {
-                UIAccessibility.post(notification: .layoutChanged, argument: object)
-            }
-        }
-    }
-}
